@@ -70,23 +100,26 @@
     "react-router-dom": "^6.0.0-beta.0",
     "react-scripts": "^4.0.3",
     "react-syntax-highlighter": "^15.4.4",
+    "stylis": "^4.0.13",
     "stylis-plugin-rtl": "^2.1.0",
     "yup": "^0.32.9"
   },
   "devDependencies": {
     "@types/react-router": "^5.1.17",
     "@typescript-eslint/eslint-plugin": "^5.2.0",
-    "eslint": "<8.0.0",
+    "eslint-config-airbnb": "^19.0.4",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.25.2",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-react": "^7.26.1",
-    "eslint-plugin-react-hooks": "^4.2.0",
+    "eslint-config-react-app": "^7.0.0",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-jsx-a11y": "^6.1.2",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-react": "^7.12.3",
+    "eslint-plugin-react-hooks": "^4.3.0",
     "husky": "^7.0.4",
     "lint-staged": "^11.2.6",
     "prettier": "^2.4.1",
     "prettier-plugin-packagejson": "^2.2.13",
     "prettier-plugin-sh": "^0.7.1",
-    "typescript": "^4.2.4"
+    "typescript": "^4.3.5"
   }
 }
