@@ -1,28 +1,22 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
-
-import { useEffect, useState } from "react";
-import type { FC } from "react";
-import { Helmet } from "react-helmet-async";
+import React from 'react';
+import type { FC } from 'react';
+import { Helmet } from 'react-helmet-async';
 import {
   Avatar,
   Box,
-  Button,
   Card,
-  CardActions,
   CardContent,
   CardHeader,
   Container,
   Grid,
   IconButton,
-  Typography,
-} from "@material-ui/core";
-import MoreVertIcon from "@material-ui/icons/MoreVert";
-import { blue } from "@material-ui/core/colors";
-import { makeStyles } from "@material-ui/core/styles";
-import TableChartIcon from "@mui/icons-material/TableChart";
-import useSettings from "../../hooks/useSettings";
-import WaterfallChartIcon from "@mui/icons-material/WaterfallChart";
-import SpectrogramTable from "src/components/plots/SpectrogramTable";
+} from '@material-ui/core';
+import MoreVertIcon from '@material-ui/icons/MoreVert';
+import { blue } from '@material-ui/core/colors';
+import { makeStyles } from '@material-ui/core/styles';
+import WaterfallChartIcon from '@mui/icons-material/WaterfallChart';
+import SpectrogramTable from 'src/components/plots/SpectrogramTable';
+import useSettings from '../../hooks/useSettings';
 
 const { REACT_APP_WS } = process.env;
 const phaseAPI = `${REACT_APP_WS}/consumer/phase`;
@@ -34,17 +28,17 @@ const useStyles = makeStyles((theme) => ({
   },
   media: {
     height: 0,
-    paddingTop: "56.25%", // 16:9
+    paddingTop: '56.25%',
   },
   expand: {
-    transform: "rotate(0deg)",
-    marginLeft: "auto",
-    transition: theme.transitions.create("transform", {
+    transform: 'rotate(0deg)',
+    marginLeft: 'auto',
+    transition: theme.transitions.create('transform', {
       duration: theme.transitions.duration.shortest,
     }),
   },
   expandOpen: {
-    transform: "rotate(180deg)",
+    transform: 'rotate(180deg)',
   },
   avatar: {
     backgroundColor: blue[500],
@@ -58,43 +52,42 @@ const PhaseDisplay: FC = () => {
   const { settings } = useSettings();
   const classes = useStyles();
 
-  console.log("PhaseDisplay: phaseAPI = ", phaseAPI);
+  console.log('PhaseDisplay: phaseAPI = ', phaseAPI);
 
-  const [data, setData] = useState(null);
-  const [socketStatus, setSocketStatus] = useState(Date().toLocaleString());
+  const [data, setData] = React.useState(null);
+  const [socketStatus, setSocketStatus] = React.useState(Date().toLocaleString());
 
-  const spectrogramTable = new SpectrogramTable("phase-display-table");
+  const spectrogramTable = new SpectrogramTable('phase-display-table');
 
   const onMessage = (event) => {
     const payload = JSON.parse(event.data);
-    console.log("PhaseDisplay:onMessage: received event.data = ", event.data, typeof event.data);
-    console.log("PhaseDisplay:onMessage: received event.data = ", payload);
+    console.log('PhaseDisplay:onMessage: received event.data = ', event.data, typeof event.data);
+    console.log('PhaseDisplay:onMessage: received event.data = ', payload);
 
-    if ("status" in payload) {
+    if ('status' in payload) {
       console.log(payload.status);
       setSocketStatus(payload.status);
     }
 
-    if ("body" in payload) {
+    if ('body' in payload) {
       setData(payload.body);
       setSocketStatus(payload.timestamp);
       spectrogramTable.draw(payload.body);
     }
   };
 
-  useEffect(() => {
-    console.log("PhaseDisplay: useEffect: 1");
+  React.useEffect(() => {
+    console.log('PhaseDisplay: useEffect: 1');
 
     ws.onmessage = onMessage;
 
     return () => {
-      // TODO
-      // ws.close();
+      // TODO :  ws.close();
     };
-  }, []);
+  }, [onMessage]);
 
-  useEffect(() => {
-    console.log("PhaseDisplay: useEffect: 2");
+  React.useEffect(() => {
+    console.log('PhaseDisplay: useEffect: 2');
     // console.log("PhaseDisplay: data = ", JSON.stringify(data));
   }, [data, socketStatus]);
 
@@ -106,12 +99,12 @@ const PhaseDisplay: FC = () => {
 
       <Box
         sx={{
-          backgroundColor: "background.default",
-          minHeight: "100%",
-          py: 8,
+          backgroundColor: 'background.default',
+          minHeight: '100%',
+          py: 3,
         }}
       >
-        <Container maxWidth={settings.compact ? "xl" : false}>
+        <Container maxWidth={settings.compact ? 'xl' : false}>
           <Grid container spacing={3}>
             <Grid item xs={12}>
               <Card>
@@ -130,8 +123,8 @@ const PhaseDisplay: FC = () => {
                   subheader={socketStatus}
                 />
 
-                <CardContent sx={{ pt: "8px" }}>
-                  <div id="phase-display-table"></div>
+                <CardContent sx={{ pt: '8px' }}>
+                  <div id="phase-display-table" />
                 </CardContent>
               </Card>
             </Grid>
