@@ -1,16 +1,17 @@
-import type { FC } from "react";
-import { useRoutes } from "react-router-dom";
-import { Toaster } from "react-hot-toast";
-import { CssBaseline, ThemeProvider } from "@material-ui/core";
+import React from 'react';
+import type { FC } from 'react';
+import { useRoutes } from 'react-router-dom';
+import { Toaster } from 'react-hot-toast';
+import { CssBaseline, ThemeProvider } from '@material-ui/core';
 
-import "./i18n";
-import SettingsDrawer from "./components/SettingsDrawer";
-import SplashScreen from "./components/SplashScreen";
-import useAuth from "./hooks/useAuth";
-import useScrollReset from "./hooks/useScrollReset";
-import useSettings from "./hooks/useSettings";
-import routes from "./routes";
-import { createCustomTheme } from "./theme";
+import './i18n';
+import SettingsDrawer from './components/SettingsDrawer';
+import SplashScreen from './components/SplashScreen';
+import useAuth from './hooks/useAuth';
+import useScrollReset from './hooks/useScrollReset';
+import useSettings from './hooks/useSettings';
+import routes from './routes';
+import { createCustomTheme } from './theme';
 
 const App: FC = () => {
   const content = useRoutes(routes);
@@ -28,10 +29,10 @@ const App: FC = () => {
 
   return (
     <ThemeProvider theme={theme}>
-        <CssBaseline />
-        <Toaster position="top-center" />
-        <SettingsDrawer />
-        {auth.isInitialized ? content : <SplashScreen />}
+      <CssBaseline />
+      <Toaster position="top-center" />
+      <SettingsDrawer />
+      {auth.isInitialized ? content : <SplashScreen />}
     </ThemeProvider>
   );
 };
