@@ -1,5 +1,5 @@
+import React from 'react';
 import type { FC, ReactNode } from 'react';
-import { useState } from 'react';
 import { Navigate, useLocation } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import useAuth from '../hooks/useAuth';
@@ -13,7 +13,7 @@ const AuthGuard: FC<AuthGuardProps> = (props) => {
   const { children } = props;
   const auth = useAuth();
   const location = useLocation();
-  const [requestedLocation, setRequestedLocation] = useState(null);
+  const [requestedLocation, setRequestedLocation] = React.useState(null);
 
   if (!auth.isAuthenticated) {
     if (location.pathname !== requestedLocation) {
@@ -35,7 +35,7 @@ const AuthGuard: FC<AuthGuardProps> = (props) => {
 };
 
 AuthGuard.propTypes = {
-  children: PropTypes.node
+  children: PropTypes.node,
 };
 
 export default AuthGuard;
