@@ -1,36 +1,11 @@
-import type { FC } from "react";
-import {
-  Box,
-  Container,
-  Divider,
-  Grid,
-  Link,
-  List,
-  ListItem,
-  ListItemAvatar,
-  ListItemText,
-  Typography,
-} from "@material-ui/core";
-import { alpha } from "@material-ui/core/styles";
-import RemoveIcon from "@material-ui/icons/Remove";
-import Logo from "./Logo";
-
-const sections = [
-  {
-    title: "Placeholders",
-    links: [
-      {
-        title: "Contact",
-        href: "#",
-      },
-    ],
-  },
-];
+import React from 'react';
+import type { FC } from 'react';
+import { Box, Container, Typography } from '@material-ui/core';
 
 const Footer: FC = (props) => (
   <Box
     sx={{
-      backgroundColor: "background.default",
+      backgroundColor: 'background.default',
       pb: 6,
       pt: {
         md: 15,
