@@ -1,13 +1,9 @@
-import type { FC, ReactNode } from "react";
-import { Navigate } from "react-router-dom";
-import PropTypes from "prop-types";
-import useAuth from "../hooks/useAuth";
+import React from 'react';
+import { Navigate } from 'react-router-dom';
+import PropTypes from 'prop-types';
+import useAuth from '../hooks/useAuth';
 
-interface GuestGuardProps {
-  children: ReactNode;
-}
-
-const GuestGuard: FC<GuestGuardProps> = ({ children }) => {
+function GuestGuard({ children }) {
   const { isAuthenticated } = useAuth();
 
   if (isAuthenticated) {
@@ -15,7 +11,7 @@ const GuestGuard: FC<GuestGuardProps> = ({ children }) => {
   }
 
   return <>{children}</>;
-};
+}
 
 GuestGuard.propTypes = {
   children: PropTypes.node,
