@@ -1,11 +1,11 @@
-import { useState } from "react";
-import type { FC, ReactNode } from "react";
-import { NavLink as RouterLink } from "react-router-dom";
-import PropTypes from "prop-types";
-import { Box, Button, Collapse, ListItem } from "@material-ui/core";
-import type { ListItemProps } from "@material-ui/core";
-import ChevronRightIcon from "@material-ui/icons/ChevronRight";
-import KeyboardArrowDownIcon from "@material-ui/icons/KeyboardArrowDown";
+import React from 'react';
+import type { FC, ReactNode } from 'react';
+import { NavLink as RouterLink } from 'react-router-dom';
+import PropTypes from 'prop-types';
+import { Box, Button, Collapse, ListItem } from '@material-ui/core';
+import type { ListItemProps } from '@material-ui/core';
+import ChevronRightIcon from '@material-ui/icons/ChevronRight';
+import KeyboardArrowDownIcon from '@material-ui/icons/KeyboardArrowDown';
 
 interface NavItemProps extends ListItemProps {
   active?: boolean;
@@ -19,18 +19,8 @@ interface NavItemProps extends ListItemProps {
 }
 
 const NavItem: FC<NavItemProps> = (props) => {
-  const {
-    active,
-    children,
-    depth,
-    icon,
-    info,
-    open: openProp,
-    path,
-    title,
-    ...other
-  } = props;
-  const [open, setOpen] = useState<boolean>(openProp);
+  const { active, children, depth, icon, info, open: openProp, path, title, ...other } = props;
+  const [open, setOpen] = React.useState<boolean>(openProp);
 
   const handleToggle = (): void => {
     setOpen((prevOpen) => !prevOpen);
@@ -48,7 +38,7 @@ const NavItem: FC<NavItemProps> = (props) => {
       <ListItem
         disableGutters
         sx={{
-          display: "block",
+          display: 'block',
           py: 0,
         }}
         {...other}
@@ -64,15 +54,15 @@ const NavItem: FC<NavItemProps> = (props) => {
           onClick={handleToggle}
           startIcon={icon}
           sx={{
-            color: "text.secondary",
-            fontWeight: "fontWeightMedium",
-            justifyContent: "flex-start",
+            color: 'text.secondary',
+            fontWeight: 'fontWeightMedium',
+            justifyContent: 'flex-start',
             pl: `${paddingLeft}px`,
-            pr: "8px",
-            py: "12px",
-            textAlign: "left",
-            textTransform: "none",
-            width: "100%",
+            pr: '8px',
+            py: '12px',
+            textAlign: 'left',
+            textTransform: 'none',
+            width: '100%',
           }}
           variant="text"
         >
@@ -89,7 +79,7 @@ const NavItem: FC<NavItemProps> = (props) => {
     <ListItem
       disableGutters
       sx={{
-        display: "flex",
+        display: 'flex',
         py: 0,
       }}
     >
@@ -97,20 +87,20 @@ const NavItem: FC<NavItemProps> = (props) => {
         component={path && RouterLink}
         startIcon={icon}
         sx={{
-          color: "text.secondary",
-          fontWeight: "fontWeightMedium",
-          justifyContent: "flex-start",
-          textAlign: "left",
+          color: 'text.secondary',
+          fontWeight: 'fontWeightMedium',
+          justifyContent: 'flex-start',
+          textAlign: 'left',
           pl: `${paddingLeft}px`,
-          pr: "8px",
-          py: "12px",
-          textTransform: "none",
-          width: "100%",
+          pr: '8px',
+          py: '12px',
+          textTransform: 'none',
+          width: '100%',
           ...(active && {
-            color: "primary.main",
-            fontWeight: "fontWeightBold",
-            "& svg": {
-              color: "primary.main",
+            color: 'primary.main',
+            fontWeight: 'fontWeightBold',
+            '& svg': {
+              color: 'primary.main',
             },
           }),
         }}
