@@ -1,41 +1,37 @@
-import { forwardRef } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import PerfectScrollbar from 'react-perfect-scrollbar';
 import type { ScrollBarProps as PerfectScrollbarProps } from 'react-perfect-scrollbar';
 import { Box } from '@material-ui/core';
 
-type ScrollbarProps = PerfectScrollbarProps
+type ScrollbarProps = PerfectScrollbarProps;
 
 const fun = (props, ref) => {
-    const { children, ...other } = props;
-  
-    const isMobile = /Android|webOS|iPhone|iPad|iPod|BlackBerry|IEMobile|Opera Mini/i.test(navigator.userAgent);
-  
-    if (isMobile) {
-      return (
-        <Box
-          ref={ref}
-          sx={{ overflowX: 'auto' }}
-        >
-          {children}
-        </Box>
-      );
-    }
-  
+  const { children, ...other } = props;
+
+  const isMobile = /Android|webOS|iPhone|iPad|iPod|BlackBerry|IEMobile|Opera Mini/i.test(
+    navigator.userAgent
+  );
+
+  if (isMobile) {
     return (
-      <PerfectScrollbar
-        ref={ref}
-        {...other}
-      >
+      <Box ref={ref} sx={{ overflowX: 'auto' }}>
         {children}
-      </PerfectScrollbar>
+      </Box>
     );
   }
-  
-const Scrollbar = forwardRef<HTMLDivElement, ScrollbarProps>(fun);
+
+  return (
+    <PerfectScrollbar ref={ref} {...other}>
+      {children}
+    </PerfectScrollbar>
+  );
+};
+
+const Scrollbar = React.forwardRef<HTMLDivElement, ScrollbarProps>(fun);
 
 Scrollbar.propTypes = {
-  children: PropTypes.node
+  children: PropTypes.node,
 };
 
 export default Scrollbar;
