@@ -1,5 +1,5 @@
-import { useEffect, useState } from "react";
-import type { FC } from "react";
+import React from 'react';
+import type { FC } from 'react';
 import {
   Box,
   Button,
@@ -10,10 +10,10 @@ import {
   TextField,
   Tooltip,
   Typography,
-} from "@material-ui/core";
-import { THEMES } from "../constants";
-import useSettings from "../hooks/useSettings";
-import MoreVertIcon from "@material-ui/icons/MoreVert";
+} from '@material-ui/core';
+import MoreVertIcon from '@material-ui/icons/MoreVert';
+import { THEMES } from '../constants';
+import useSettings from '../hooks/useSettings';
 
 const getValues = (settings) => ({
   compact: settings.compact,
@@ -25,10 +25,10 @@ const getValues = (settings) => ({
 
 const SettingsDrawer: FC = () => {
   const { settings, saveSettings } = useSettings();
-  const [open, setOpen] = useState<boolean>(false);
-  const [values, setValues] = useState(getValues(settings));
+  const [open, setOpen] = React.useState<boolean>(false);
+  const [values, setValues] = React.useState(getValues(settings));
 
-  useEffect(() => {
+  React.useEffect(() => {
     setValues(getValues(settings));
   }, [settings]);
 
@@ -62,7 +62,7 @@ const SettingsDrawer: FC = () => {
           sx={{
             bottom: 0,
             margin: (theme) => theme.spacing(4),
-            position: "fixed",
+            position: 'fixed',
             right: 0,
             zIndex: (theme) => theme.zIndex.speedDial,
           }}
@@ -89,9 +89,7 @@ const SettingsDrawer: FC = () => {
             fullWidth
             label="Theme"
             name="theme"
-            onChange={(event): void =>
-              handleChange("theme", event.target.value)
-            }
+            onChange={(event): void => handleChange('theme', event.target.value)}
             select
             SelectProps={{ native: true }}
             value={values.theme}
@@ -100,14 +98,14 @@ const SettingsDrawer: FC = () => {
             {Object.keys(THEMES).map((theme) => (
               <option key={theme} value={theme}>
                 {theme
-                  .split("_")
+                  .split('_')
                   .map((w) => w[0].toUpperCase() + w.substr(1).toLowerCase())
-                  .join(" ")}
+                  .join(' ')}
               </option>
             ))}
           </TextField>
         </Box>
-         
+
         <Box
           sx={{
             mt: 2,
@@ -122,18 +120,14 @@ const SettingsDrawer: FC = () => {
                 edge="start"
                 name="direction"
                 onChange={(event): void =>
-                  handleChange("responsiveFontSizes", event.target.checked)
+                  handleChange('responsiveFontSizes', event.target.checked)
                 }
               />
             }
             label={
               <div>
                 Responsive font sizes
-                <Typography
-                  color="textSecondary"
-                  component="p"
-                  variant="caption"
-                >
+                <Typography color="textSecondary" component="p" variant="caption">
                   Adjust font for small devices
                 </Typography>
               </div>
@@ -153,19 +147,13 @@ const SettingsDrawer: FC = () => {
                 color="primary"
                 edge="start"
                 name="compact"
-                onChange={(event): void =>
-                  handleChange("compact", event.target.checked)
-                }
+                onChange={(event): void => handleChange('compact', event.target.checked)}
               />
             }
             label={
               <div>
                 Compact
-                <Typography
-                  color="textSecondary"
-                  component="p"
-                  variant="caption"
-                >
+                <Typography color="textSecondary" component="p" variant="caption">
                   Fixed width on some screens
                 </Typography>
               </div>
@@ -185,19 +173,13 @@ const SettingsDrawer: FC = () => {
                 color="primary"
                 edge="start"
                 name="roundedCorners"
-                onChange={(event): void =>
-                  handleChange("roundedCorners", event.target.checked)
-                }
+                onChange={(event): void => handleChange('roundedCorners', event.target.checked)}
               />
             }
             label={
               <div>
                 Rounded Corners
-                <Typography
-                  color="textSecondary"
-                  component="p"
-                  variant="caption"
-                >
+                <Typography color="textSecondary" component="p" variant="caption">
                   Increase border radius
                 </Typography>
               </div>
@@ -205,12 +187,7 @@ const SettingsDrawer: FC = () => {
           />
         </Box>
         <Box sx={{ mt: 3 }}>
-          <Button
-            color="primary"
-            fullWidth
-            onClick={handleSave}
-            variant="contained"
-          >
+          <Button color="primary" fullWidth onClick={handleSave} variant="contained">
             Save Settings
           </Button>
         </Box>
