@@ -1,26 +1,41 @@
-import * as d3 from "d3";
-import * as _ from "lodash";
-import Spectrogram from "./Spectrogram";
-import "./SpectrogramTable.css";
+/* eslint-disable @typescript-eslint/no-explicit-any */
+/* eslint-disable @typescript-eslint/no-unused-vars */
+import * as d3 from 'd3';
+import * as _ from 'lodash';
+import Spectrogram from './Spectrogram';
+import './SpectrogramTable.css';
+
+type NewType = any;
 
 class SpectrogramTable {
   phaseDisplayId;
+
   table;
+
   cells;
+
   colHeaders;
+
   rowHeaders;
+
   numRows;
+
   numCols;
+
   cellData;
 
-  constructor(id) {
+  constructor(id: string) {
     this.phaseDisplayId = id;
   }
 
-  draw(data) {
+  draw(data: NewType): void {
     this.cellData = data.phase_values;
 
-    if (!this.table || !_.isEqual(this.colHeaders, data.polarisation) || !_.isEqual(this.rowHeaders, data.baseline)) {
+    if (
+      !this.table ||
+      !_.isEqual(this.colHeaders, data.polarisation) ||
+      !_.isEqual(this.rowHeaders, data.baseline)
+    ) {
       this.colHeaders = data.polarisation;
       this.rowHeaders = data.baseline;
       this.numCols = this.colHeaders.length;
@@ -46,67 +61,61 @@ class SpectrogramTable {
     }
   }
 
-  drawTable() {
+  drawTable(): void {
     // remove existing table
-    d3.select("#" + this.phaseDisplayId)
-      .selectAll("table")
-      .remove();
+    d3.select(`#${this.phaseDisplayId}`).selectAll('table').remove();
 
-    const table = d3
-      .select("#" + this.phaseDisplayId)
-      .append("table")
-      .style("class", "table");
+    const table = d3.select(`#${this.phaseDisplayId}`).append('table').style('class', 'table');
 
-    const thead = table.append("thead");
-    const tbody = table.append("tbody");
+    const thead = table.append('thead');
+    const tbody = table.append('tbody');
 
     // append the column header row
     thead
-      .append("tr")
-      .selectAll("th")
-      .data([""].concat(this.colHeaders)) // column "" is for row header
+      .append('tr')
+      .selectAll('th')
+      .data([''].concat(this.colHeaders)) // column "" is for row header
       .enter()
-      .append("th")
+      .append('th')
       .text((col) => col)
-      .style("class", "th");
+      .style('class', 'th');
 
     // create a row for each object/array in the data
     const rows = tbody
-      .selectAll("tr")
+      .selectAll('tr')
       .data(this.cells)
       .enter()
-      .append("tr")
-      .attr("id", (d, i) => {
-        return i;
-      });
+      .append('tr')
+      .attr('id', (d, i) => i);
 
     // create row header
     rows
-      .append("th")
-      .attr("scope", "row")
+      .append('th')
+      .attr('scope', 'row')
       .text((row, i) => this.rowHeaders[i]);
 
     // create a cell in each row for each column
     rows
-      .selectAll("td")
-      .data((row, i) => {
-        //
-        return this.colHeaders;
-      })
+      .selectAll('td')
+      .data(
+        (row, i) =>
+          //
+          this.colHeaders
+      )
       .enter()
-      .append("td")
-      .attr("id", function (d, i) {
+      .append('td')
+      .attr('id', function (d, i) {
         // the current node is selected using 'this', hence use 'function', not '=>'
         const trId = d3.select(this).node().parentNode.id;
         return `${trId}${i}`;
       })
-      .append("canvas")
-      .attr("id", function (d, i) {
+      .append('canvas')
+      .attr('id', function (d, i) {
         // the current node is selected using 'this', hence use 'function', not '=>'
         const tdId = d3.select(this).node().parentNode.id;
         return `canvas${tdId}`;
       })
-      .attr("style", "canvas");
+      .attr('style', 'canvas');
 
     return table;
   }
