@@ -1,10 +1,10 @@
-import { useEffect } from 'react';
+import React from 'react';
 import { useLocation } from 'react-router-dom';
 
 const useScrollReset = (): null => {
   const location = useLocation();
 
-  useEffect(() => {
+  React.useEffect(() => {
     window.scrollTo(0, 0);
   }, [location.pathname]);
 
