@@ -4,23 +4,21 @@ import { initReactI18next } from 'react-i18next';
 const resources = {
   en: {
     translation: {
-      'Welcome': 'Welcome'
-    }
+      Welcome: 'Welcome',
+    },
   },
   fr: {
     translation: {
-      'Welcome': 'Bienvenue'
-    }
-  }
+      Welcome: 'Bienvenue',
+    },
+  },
 };
 
-i18n
-  .use(initReactI18next)
-  .init({
-    resources,
-    lng: 'en',
-    fallbackLng: 'en',
-    interpolation: {
-      escapeValue: false
-    }
-  });
+i18n.use(initReactI18next).init({
+  resources,
+  lng: 'en',
+  fallbackLng: 'en',
+  interpolation: {
+    escapeValue: false,
+  },
+});
