@@ -1,17 +1,17 @@
-import "react-perfect-scrollbar/dist/css/styles.css";
-import "react-draft-wysiwyg/dist/react-draft-wysiwyg.css";
-import "react-quill/dist/quill.snow.css";
-import "nprogress/nprogress.css";
-import { StrictMode } from "react";
-import ReactDOM from "react-dom";
-import { HelmetProvider } from "react-helmet-async";
-import { BrowserRouter } from "react-router-dom";
-import LocalizationProvider from "@material-ui/lab/LocalizationProvider";
-import AdapterDateFns from "@material-ui/lab/AdapterDateFns";
-import StyledEngineProvider from "@material-ui/core/StyledEngineProvider";
-import App from "./App";
-import { AuthProvider } from "./contexts/JWTContext";
-import { SettingsProvider } from "./contexts/SettingsContext";
+import 'react-perfect-scrollbar/dist/css/styles.css';
+import 'react-draft-wysiwyg/dist/react-draft-wysiwyg.css';
+import 'react-quill/dist/quill.snow.css';
+import 'nprogress/nprogress.css';
+import React, { StrictMode } from 'react';
+import ReactDOM from 'react-dom';
+import { HelmetProvider } from 'react-helmet-async';
+import { BrowserRouter } from 'react-router-dom';
+import LocalizationProvider from '@material-ui/lab/LocalizationProvider';
+import AdapterDateFns from '@material-ui/lab/AdapterDateFns';
+import StyledEngineProvider from '@material-ui/core/StyledEngineProvider';
+import App from './App';
+import { AuthProvider } from './contexts/JWTContext';
+import { SettingsProvider } from './contexts/SettingsContext';
 
 ReactDOM.render(
   <StrictMode>
@@ -29,5 +29,5 @@ ReactDOM.render(
       </StyledEngineProvider>
     </HelmetProvider>
   </StrictMode>,
-  document.getElementById("root")
+  document.getElementById('root')
 );
