@@ -1,17 +1,12 @@
-import type { FC } from "react";
-import { Link as RouterLink } from "react-router-dom";
-import { Helmet } from "react-helmet-async";
-import {
-  Box,
-  Card,
-  CardContent,
-  Container,
-  Divider,
-  Typography,
-} from "@material-ui/core";
-import { LoginJWT } from "../../components/auth/login";
-import Logo from "../../components/Logo";
-import useAuth from "../../hooks/useAuth";
+/* eslint-disable @typescript-eslint/no-explicit-any */
+import React from 'react';
+import type { FC } from 'react';
+import { Link as RouterLink } from 'react-router-dom';
+import { Helmet } from 'react-helmet-async';
+import { Box, Card, CardContent, Container, Divider, Typography } from '@material-ui/core';
+import { LoginJWT } from '../../components/auth/login';
+import Logo from '../../components/Logo';
+import useAuth from '../../hooks/useAuth';
 
 const Login: FC = () => {
   const { platform } = useAuth() as any;
@@ -23,17 +18,17 @@ const Login: FC = () => {
       </Helmet>
       <Box
         sx={{
-          backgroundColor: "background.default",
-          display: "flex",
-          flexDirection: "column",
-          minHeight: "100vh",
+          backgroundColor: 'background.default',
+          display: 'flex',
+          flexDirection: 'column',
+          minHeight: '100vh',
         }}
       >
-        <Container maxWidth="sm" sx={{ py: "80px" }}>
+        <Container maxWidth="sm" sx={{ py: '80px' }}>
           <Box
             sx={{
-              display: "flex",
-              justifyContent: "center",
+              display: 'flex',
+              justifyContent: 'center',
               mb: 8,
             }}
           >
@@ -49,16 +44,16 @@ const Login: FC = () => {
           <Card>
             <CardContent
               sx={{
-                display: "flex",
-                flexDirection: "column",
+                display: 'flex',
+                flexDirection: 'column',
                 p: 4,
               }}
             >
               <Box
                 sx={{
-                  alignItems: "center",
-                  display: "flex",
-                  justifyContent: "space-between",
+                  alignItems: 'center',
+                  display: 'flex',
+                  justifyContent: 'space-between',
                   mb: 3,
                 }}
               >
@@ -77,7 +72,7 @@ const Login: FC = () => {
                   mt: 3,
                 }}
               >
-                {platform === "JWT" && <LoginJWT />}
+                {platform === 'JWT' && <LoginJWT />}
               </Box>
               <Divider sx={{ my: 3 }} />
             </CardContent>
