@@ -1,25 +1,23 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 
-import { FC, useEffect } from "react";
-import { Helmet } from "react-helmet-async";
+import React, { FC } from 'react';
+import { Helmet } from 'react-helmet-async';
 import {
   Avatar,
   Box,
   Card,
-  CardActions,
   CardContent,
   CardHeader,
   Container,
   Grid,
   IconButton,
-  Typography,
-} from "@material-ui/core";
-import MoreVertIcon from "@material-ui/icons/MoreVert";
-import DashboardIcon from "@material-ui/icons/Dashboard";
-import { makeStyles } from "@material-ui/core/styles";
-import { blue } from "@material-ui/core/colors";
+} from '@material-ui/core';
+import MoreVertIcon from '@material-ui/icons/MoreVert';
+import DashboardIcon from '@material-ui/icons/Dashboard';
+import { makeStyles } from '@material-ui/core/styles';
+import { blue } from '@material-ui/core/colors';
 
-import useSettings from "../../hooks/useSettings";
+import useSettings from '../../hooks/useSettings';
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -27,17 +25,17 @@ const useStyles = makeStyles((theme) => ({
   },
   media: {
     height: 0,
-    paddingTop: "56.25%", // 16:9
+    paddingTop: '56.25%', // 16:9
   },
   expand: {
-    transform: "rotate(0deg)",
-    marginLeft: "auto",
-    transition: theme.transitions.create("transform", {
+    transform: 'rotate(0deg)',
+    marginLeft: 'auto',
+    transition: theme.transitions.create('transform', {
       duration: theme.transitions.duration.shortest,
     }),
   },
   expandOpen: {
-    transform: "rotate(180deg)",
+    transform: 'rotate(180deg)',
   },
   avatar: {
     backgroundColor: blue[500],
@@ -59,12 +57,12 @@ const Overview: FC = () => {
 
       <Box
         sx={{
-          backgroundColor: "background.default",
-          minHeight: "100%",
-          py: 8,
+          backgroundColor: 'background.default',
+          minHeight: '100%',
+          py: 3,
         }}
       >
-        <Container maxWidth={settings.compact ? "xl" : false}>
+        <Container maxWidth={settings.compact ? 'xl' : false}>
           <Grid container spacing={3}>
             <Grid item xs={12}>
               <Card>
@@ -83,7 +81,7 @@ const Overview: FC = () => {
                   subheader=""
                 />
 
-                <CardContent sx={{ pt: "8px" }}>
+                <CardContent sx={{ pt: '8px' }}>
                   <h3>This is an Overview Dashboard for the QA Display</h3>
                   <p>This page is under construction</p>
                 </CardContent>
