@@ -1,30 +1,25 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
-
-import { useEffect, useState } from "react";
-import type { FC } from "react";
-import { Helmet } from "react-helmet-async";
+import React from 'react';
+import type { FC } from 'react';
+import { Helmet } from 'react-helmet-async';
 import {
   Avatar,
   Box,
-  Button,
   Card,
-  CardActions,
   CardContent,
   CardHeader,
   Container,
   Grid,
   IconButton,
-  Typography,
-} from "@material-ui/core";
-import MoreVertIcon from "@material-ui/icons/MoreVert";
-import { blue } from "@material-ui/core/colors";
-import { makeStyles } from "@material-ui/core/styles";
-import TimelineIcon from "@material-ui/icons/Timeline";
+} from '@material-ui/core';
+import MoreVertIcon from '@material-ui/icons/MoreVert';
+import { blue } from '@material-ui/core/colors';
+import { makeStyles } from '@material-ui/core/styles';
+import TimelineIcon from '@material-ui/icons/Timeline';
+import useSettings from '../../hooks/useSettings';
+import Spectrum from '../../components/plots/Spectrum';
+import mockSpectrumDataPayload from '../../mock/mockSpectrumData';
 
-import useSettings from "../../hooks/useSettings";
-import Spectrum from "../../components/plots/Spectrum";
-import mockSpectrumDataPayload from "../../mock/mockSpectrumData";
-
+const MOCK_DATA = false;
 const { REACT_APP_WS } = process.env;
 const spectrumAPI = `${REACT_APP_WS}/consumer/spectrum`;
 const ws = new WebSocket(spectrumAPI);
@@ -35,17 +30,17 @@ const useStyles = makeStyles((theme) => ({
   },
   media: {
     height: 0,
-    paddingTop: "56.25%", // 16:9
+    paddingTop: '56.25%',
   },
   expand: {
-    transform: "rotate(0deg)",
-    marginLeft: "auto",
-    transition: theme.transitions.create("transform", {
+    transform: 'rotate(0deg)',
+    marginLeft: 'auto',
+    transition: theme.transitions.create('transform', {
       duration: theme.transitions.duration.shortest,
     }),
   },
   expandOpen: {
-    transform: "rotate(180deg)",
+    transform: 'rotate(180deg)',
   },
   avatar: {
     backgroundColor: blue[500],
@@ -59,38 +54,37 @@ const SpectrumPlot: FC = () => {
   const { settings } = useSettings();
   const classes = useStyles();
 
-  console.log("App: spectrumAPI: ", spectrumAPI);
+  console.log('App: spectrumAPI: ', spectrumAPI);
 
-  const [data, setData] = useState([]); // mockSpectrumDataPayload.body
-  const [socketStatus, setSocketStatus] = useState(Date().toLocaleString());
+  const [data, setData] = React.useState(MOCK_DATA ? mockSpectrumDataPayload.body : []);
+  const [socketStatus, setSocketStatus] = React.useState(Date().toLocaleString());
 
   const onMessage = (event) => {
     const payload = JSON.parse(event.data);
-    console.log("App.js:onMessage: received payload = ", payload);
+    console.log('App.js:onMessage: received payload = ', payload);
 
-    if ("status" in payload) {
+    if ('status' in payload) {
       console.log(payload.status);
       setSocketStatus(payload.status);
     }
 
-    if ("body" in payload) {
+    if ('body' in payload) {
       setData(payload.body);
       setSocketStatus(payload.timestamp);
     }
   };
 
-  useEffect(() => {
-    console.log("App: useEffect: 1");
+  React.useEffect(() => {
+    console.log('App: useEffect: 1');
     ws.onmessage = onMessage;
 
     return () => {
-      // TODO
-      // ws.close();
+      // TODO - ws.close();
     };
   });
 
-  useEffect(() => {
-    console.log("App: useEffect: 2");
+  React.useEffect(() => {
+    console.log('App: useEffect: 2');
   }, [data, socketStatus]);
 
   return (
@@ -101,12 +95,12 @@ const SpectrumPlot: FC = () => {
 
       <Box
         sx={{
-          backgroundColor: "background.default",
-          minHeight: "100%",
-          py: 8,
+          backgroundColor: 'background.default',
+          minHeight: '100%',
+          py: 3,
         }}
       >
-        <Container maxWidth={settings.compact ? "xl" : false}>
+        <Container maxWidth={settings.compact ? 'xl' : false}>
           <Grid container spacing={3}>
             <Grid item xs={12}>
               <Card>
@@ -125,7 +119,7 @@ const SpectrumPlot: FC = () => {
                   subheader={socketStatus}
                 />
 
-                <CardContent sx={{ pt: "8px" }}>
+                <CardContent sx={{ pt: '8px' }}>
                   <Spectrum width={1000} height={500} data={data} />
                 </CardContent>
               </Card>
