@@ -1,38 +1,39 @@
-import { Suspense, lazy } from "react";
-import type { PartialRouteObject } from "react-router";
-import { Navigate } from "react-router-dom";
-import AuthGuard from "./components/AuthGuard";
+import React, { Suspense, lazy } from 'react';
+import type { RouteObject } from 'react-router';
+import { Navigate } from 'react-router-dom';
+import AuthGuard from './components/AuthGuard';
 
-import GuestGuard from "./components/GuestGuard";
-import DashboardLayout from "./components/dashboard/DashboardLayout";
-import LoadingScreen from "./components/LoadingScreen";
-import MainLayout from "./components/MainLayout";
+import GuestGuard from './components/GuestGuard';
+import DashboardLayout from './components/dashboard/DashboardLayout';
+import LoadingScreen from './components/LoadingScreen';
+import MainLayout from './components/MainLayout';
 
-// eslint-disable-next-line react/display-name
-const Loadable = (CustomComponent) => (props) =>
-  (
-    <Suspense fallback={<LoadingScreen />}>
-      <CustomComponent {...props} />
-    </Suspense>
-  );
+const Loadable = (CustomComponent) =>
+  function (props) {
+    return (
+      <Suspense fallback={<LoadingScreen />}>
+        <CustomComponent {...props} />
+      </Suspense>
+    );
+  };
 
 // Authentication pages
-const Login = Loadable(lazy(() => import("./pages/auth/Login")));
+const Login = Loadable(lazy(() => import('./pages/auth/Login')));
 
 // Dashboard & plot pages
-const Overview = Loadable(lazy(() => import("./pages/dashboards/Overview")));
-const SpectrumPlot = Loadable(lazy(() => import("./pages/plots/SpectrumPlot")));
-const PhaseDisplay = Loadable(lazy(() => import("./pages/plots/PhaseDisplay")));
-const RfiDisplay = Loadable(lazy(() => import("./pages/plots/RfiDisplay")));
+const Overview = Loadable(lazy(() => import('./pages/dashboards/Overview')));
+const SpectrumPlot = Loadable(lazy(() => import('./pages/plots/SpectrumPlot')));
+const PhaseDisplay = Loadable(lazy(() => import('./pages/plots/PhaseDisplay')));
+const RfiDisplay = Loadable(lazy(() => import('./pages/plots/RfiDisplay')));
 
 // Error pages
-const AuthorizationRequired = Loadable(lazy(() => import("./pages/AuthorizationRequired")));
-const NotFound = Loadable(lazy(() => import("./pages/NotFound")));
-const ServerError = Loadable(lazy(() => import("./pages/ServerError")));
+const AuthorizationRequired = Loadable(lazy(() => import('./pages/AuthorizationRequired')));
+const NotFound = Loadable(lazy(() => import('./pages/NotFound')));
+const ServerError = Loadable(lazy(() => import('./pages/ServerError')));
 
-const routes: PartialRouteObject[] = [
+const routes: RouteObject[] = [
   {
-    path: "dashboard",
+    path: 'dashboard',
     element: (
       <AuthGuard>
         <DashboardLayout />
@@ -40,17 +41,17 @@ const routes: PartialRouteObject[] = [
     ),
     children: [
       {
-        path: "/",
+        path: '',
         element: <Navigate to="/dashboard/overview" replace />,
       },
       {
-        path: "overview",
+        path: 'overview',
         element: <Overview />,
       },
     ],
   },
   {
-    path: "plot",
+    path: 'plot',
     element: (
       <AuthGuard>
         <DashboardLayout />
@@ -58,28 +59,28 @@ const routes: PartialRouteObject[] = [
     ),
     children: [
       {
-        path: "/",
+        path: '',
         element: <Navigate to="/plot/spectrum-plot" replace />,
       },
       {
-        path: "spectrum-plot",
+        path: 'spectrum-plot',
         element: <SpectrumPlot />,
       },
       {
-        path: "phase-display",
+        path: 'phase-display',
         element: <PhaseDisplay />,
       },
       {
-        path: "rfi-display",
+        path: 'rfi-display',
         element: <RfiDisplay />,
       },
     ],
   },
   {
-    path: "auth",
+    path: 'auth',
     children: [
       {
-        path: "login",
+        path: 'login',
         element: (
           <GuestGuard>
             <Login />
@@ -89,27 +90,27 @@ const routes: PartialRouteObject[] = [
     ],
   },
   {
-    path: "*",
+    path: '*',
     element: <MainLayout />,
     children: [
       {
-        path: "/",
+        path: '',
         element: <Navigate to="/dashboard/overview" replace />,
       },
       {
-        path: "401",
+        path: '401',
         element: <AuthorizationRequired />,
       },
       {
-        path: "404",
+        path: '404',
         element: <NotFound />,
       },
       {
-        path: "500",
+        path: '500',
         element: <ServerError />,
       },
       {
-        path: "*",
+        path: '*',
         element: <NotFound />,
       },
     ],
