@@ -3,5 +3,5 @@ export interface User {
   avatar: string;
   email: string;
   name: string;
-  [key: string]: any;
+  [key: string]: string;
 }
